Add and improve test setup macros.
The two new macros kill any buffers that are created during the exection
of a given test and allow you to test in an enivronment where FlyKey is
actually running, respectively. The new macros require the common lisp
package, cl, for Emacs, so it was necessary to install it in the Cask
file. Finally, creating the new test helped me find and fix a bug which
would occur when trying to add keybindings using an empty flybuf. I also
added new tests to check the functioning of the macros and to try and
find out why using the buffer-list-update-hook to update the keybindings
isn't working.