Implement command keybindings.
Add the option to define on-the-fly bindings to Elisp commands. The
functionality to produce the command pairs is split into two
functions, flykey-create-command-binding-cmds and
flykey-create-normal-binding-cmds. > is used for commands and = is
still used for the normal bindings. There are also new tests for the
added functionality, and an update to README.